+++ sentry-android-core/src/test/java/io/sentry/android/core/SentryShadowActivityManager.kt
@@ -10,18 +10,25 @@ import org.robolectric.annotation.Implements
 class SentryShadowActivityManager {
   companion object {
     private var historicalProcessStartReasons: List<ApplicationStartInfo> = emptyList()
+    private var historicalProcessStartReasonsException: RuntimeException? = null
 
     fun setHistoricalProcessStartReasons(startReasons: List<ApplicationStartInfo>) {
       historicalProcessStartReasons = startReasons
     }
 
+    fun setHistoricalProcessStartReasonsException(exception: RuntimeException) {
+      historicalProcessStartReasonsException = exception
+    }
+
     fun reset() {
       historicalProcessStartReasons = emptyList()
+      historicalProcessStartReasonsException = null
     }
   }
 
   @Implementation
   fun getHistoricalProcessStartReasons(maxNum: Int): List<ApplicationStartInfo> {
+    historicalProcessStartReasonsException?.let { throw it }
     return historicalProcessStartReasons.take(maxNum)
   }
 }